Modify migrate-images-list subcommand to be more granular
I wrote most of this branch quite a while ago. Originally I intended
that all of the images not mentioned by retrieve-image-tags/config.json
would be migrated in one step. However there is now the need to be
more granular as to which images are migrated. The work on this branch
up to this point satisfies this requirement.