Do not mirror images to registries that they come from
In the set of images migrated by this PR there are two
images that come from the registry.suse.com/rancher registry.
This commit modifies the regsync.yaml generation code so
that we do not mirror these images back to their source
registry. This is a good principle to have in general,
I think.